+++ tests/testthat/test-sys-time.R
@@ -217,6 +217,23 @@ test_that("empty dots are checked", {
 })
 
 # ------------------------------------------------------------------------------
+# vec_ptype_full() / vec_ptype_abbr()
+
+test_that("`vec_ptype_full()` prints correctly", {
+  expect_snapshot({
+    vec_ptype_full(sys_days())
+    vec_ptype_full(sys_seconds(1:5))
+  })
+})
+
+test_that("`vec_ptype_abbr()` prints correctly", {
+  expect_snapshot({
+    vec_ptype_abbr(sys_days())
+    vec_ptype_abbr(sys_seconds(1:5))
+  })
+})
+
+# ------------------------------------------------------------------------------
 # vec_ptype()
 
 test_that("ptype is correct", {